moved vimColorParser to a separate parser file
